Refactored the code that runs the ability system.
Added AbilityCast to store all the data for an ability that is cast.  This can more easily be passed through many functions without parameter bloat. Moved as many functions as possible that were not unique to player character functionality into the Character class so that enemy AI can make use of it in the future.

IMPORTANT NOTE: I moved Health Bar Controller and Energy Bar Controller scripts back onto the characters (player, enemy, etc) in my scene. This was the best way to make it work after the refactor.